+++ sites/www/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 .. note::
     Looking for the Fabric 1.x changelog? See :doc:`/changelog-v1`.
 
+- :feature: `-` Invoke fabric package using m flag; python -m fabric.
 - :support:`1761 backported` Integration tests were never added to Travis or
   ported to pytest before 2.0's release; this has been addressed.
 - :support:`1759 backported` Apply the ``black`` code formatter to the codebase